Overbrighten backed lightmaps.

+++ code/renderer_bgfx/Main.cpp
@@ -1739,7 +1739,7 @@ void Main::setWindowGamma()
 			value = int(255 * pow(i / 255.0f, 1.0f / gamma) + 0.5f);
 		}
 
-		g_gammaTable[i] = math::Clamped(value * g_overbrightFactor, 0, 255);
+		g_gammaTable[i] = math::Clamped(int(value * g_overbrightFactor), 0, 255);
 	}
 
 	window::SetGamma(g_gammaTable, g_gammaTable, g_gammaTable);
