+++ code/renderer_bgfx/Precompiled.h
@@ -1756,7 +1756,7 @@ namespace world
 static const size_t g_funcTableSize = 1024;
 static const size_t g_funcTableSize2 = 10;
 static const size_t g_funcTableMask = g_funcTableSize - 1;
-static const int g_overbrightFactor = 2;
+static const float g_overbrightFactor = 2.0f;
 static const float g_identityLight = 1.0f / g_overbrightFactor;
 static const size_t g_gammaTableSize = 256;
